Create SECURITY.md
Access to the account is limited and all private cooperation is canceled.  Only the owner owns 10,000% of the account, changes, and codes, and the responsibility should not be transferred to another person. Please return sei currency to the protocol and leave the chain network. The chain network can be hacked by Validator.