Delete readme.md
Please change the chain network environment to local network.  The security there is safe for me.  The web3 network has many potential risks for me.  I want to have sei cryptocurrency locally.  My trust wallet has been abused.  It is used and scammers are waiting for my currency to be released on a smart chain.  For phishing I have documents to prove it. Fraudster with access to my account. Changing programs to their advantage. Restore my account to 17th December 2023 and access is cut off except for the owner.